+++ src/modules/views/index.ts
@@ -1,0 +1,6 @@
+import { getCreatorsExtOptionDialog } from "./creatorsExtDialog";
+import { duplicationDialog } from "./duplicationDialog";
+import { richTextToolBar } from "./richTextToolBar";
+import { setLanguageManualDialog } from "./setLanguageManualDialog";
+
+export { getCreatorsExtOptionDialog, richTextToolBar, duplicationDialog, setLanguageManualDialog };